refactor session module
