+++ wt_mdb/src/lib.rs
@@ -172,7 +172,7 @@ impl<'a> RecordView<'a> {
 pub type Record = RecordView<'static>;
 
 pub use connection::Connection;
-pub use session::{RecordCursor, RecordCursorGuard, Session, StatisticsCursor};
+pub use session::{RecordCursor, RecordCursorGuard, Session, StatCursor};
 pub type Result<T> = std::result::Result<T, Error>;
 
 fn make_result<T>(code: i32, value: T) -> Result<T> {
